+++ src/main/kotlin/me/fzzyhmstrs/fzzy_config/impl/ConfigApiImpl.kt
@@ -27,6 +27,7 @@ import me.fzzyhmstrs.fzzy_config.config.ConfigContext
 import me.fzzyhmstrs.fzzy_config.config.ConfigContext.Keys.ACTIONS
 import me.fzzyhmstrs.fzzy_config.config.ConfigContext.Keys.RESTART_RECORDS
 import me.fzzyhmstrs.fzzy_config.config.ConfigContext.Keys.VERSIONS
+import me.fzzyhmstrs.fzzy_config.config.ConfigSection
 import me.fzzyhmstrs.fzzy_config.entry.Entry
 import me.fzzyhmstrs.fzzy_config.entry.EntryDeserializer
 import me.fzzyhmstrs.fzzy_config.entry.EntrySerializer
@@ -53,8 +54,10 @@ import kotlin.experimental.and
 import kotlin.experimental.or
 import kotlin.math.min
 import kotlin.reflect.*
+import kotlin.reflect.full.allSuperclasses
 import kotlin.reflect.full.findAnnotation
 import kotlin.reflect.full.memberProperties
+import kotlin.reflect.full.superclasses
 import kotlin.reflect.jvm.javaField
 
 internal object ConfigApiImpl {
@@ -295,7 +298,11 @@ internal object ConfigApiImpl {
             }
             val ignoreVisibility = isIgnoreVisibility(config::class) || ignoreVisibility(flags)
             //java fields are ordered in declared order, apparently not so for Kotlin properties. use these first to get ordering. skip Transient
-            val fields = config::class.java.fields.filter { !isTransient(it.modifiers) }
+            val fields = config::class.java.declaredFields.filter { !isTransient(it.modifiers) }.toMutableList()
+            for (sup in config::class.allSuperclasses) {
+                if (sup == Config::class) continue //ignore Config itself, as that has state we don't need
+                fields.addAll(sup.java.declaredFields.filter { !isTransient(it.modifiers) })
+            }
             //generate an index map, so I can order the properties based on name
             val orderById = fields.withIndex().associate { it.value.name to it.index }
             //kotlin member properties filtered by [field map contains it && if NonSync matters, it isn't NonSync]. NonSync does not matter by default
@@ -408,7 +415,13 @@ internal object ConfigApiImpl {
             val globalAction = getAction(config::class.annotations)
             val ignoreVisibility = isIgnoreVisibility(config::class) || ignoreVisibility(flags)
 
-            val fields = config::class.java.fields.filter { !isTransient(it.modifiers) }
+            val fields = config::class.java.declaredFields.filter { !isTransient(it.modifiers) }.toMutableList()
+            for (sup in config::class.allSuperclasses) {
+                if (sup == Config::class) continue
+                fields.addAll(sup.java.declaredFields.filter { !isTransient(it.modifiers) })
+            }
+
+            println(fields.map { it.name })
             val orderById = fields.withIndex().associate { it.value.name to it.index }
             for (prop in config.javaClass.kotlin.memberProperties.filter {
                 if (ignoreNonSync(flags)) true else !isNonSync(it)